+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## [0.0.1] - TODO: Add release date.
+## [1.1.0] - Fixed some minor bugs.
 
-* TODO: Describe initial release.
+* Updated the documentation.
+* Added button mode with `onPressed`.
+* Added `mask` (clip) feature.
+* Added `update(child)` feature.
+* Fixed some minor bugs.
+* TODO: Performance improvement.
+
+## [1.0.0] - Publish the widget.
+
+* TODO: Wait for the bugs?
+* TODO: Performance improvement.